Fix off-by-one timestamps
This major commit ensures that the initial time is always 0 if none is known,
in order to prevent off-by-one errors when the timeline requests states from
the game state history.

We also replace older states if we have multiple states with the same
timestamp in order to prevent scrubbing issues in history depending on
whether we arrive from the tail or head.

Fixes #60.